Fix failures with double #
Why?

When an interest already existed and the user attempted to add the
interest and accidentally included the "#" in the name, it would fail to
create any interests for that announcement

How does this address the need?

It uses `Repo.get_or_insert` to make sure that it is checking for
existing interests without a "#"