Add unsubscribe link to emails
This adds an unsubscribe link to each users email by taking advantage of
mandrills merge_vars feature and handlebar templating.

Subscriptions are now created with a token generated and each comment
email sent out for a user with a subscription is given a url that allows
them to unsubscribe from that announcement via the subscription token.