Link new interests in daily digest
The daily digest contains new interests and new announcements. The
announcements are linked but the interests are not. Most of the time the
post that triggered the new interest is in the list of new
announcements, but it doesn't have to be.

Linking the interests as well keeps the mail consistent.