Add login and basic announcements page
This adds the login page with new OAuth callbacks to separate the
JavaScript client login and the web client login. This also adds a
very basic announcements page so we can redirect and view
announcements.

This also prefixes api urls with `api_` to prevent collisions with their
non API counterparts.