Add announcements page
The announcements page is completely functional with the exception of
the interest picker. This includes switching between "all" and "my"
announcements.

Timex was upgraded in this PR as well to 2.x to allow us to take
advantage of the relative time formatter.

We also need to sort announcements by when they were last updated via
comments or the announcement itself. The current approach is to fetch
the announcements and sort manually which is slower than building a
query to do this, but this is easier for now.