Build URI based on conn
The way elixir `*_url` helpers work is that it reads the config to get
the host/port. Since it uses the `$PORT` environment variable by default
we get url's generated pointing at the port we bind on instead of 80, or
443.

This change uses the `conn` to build a `%URI{}` which uses the requests
host and scheme instead of the configurations.