Fix large and weird looking buttons
The problem was that the old site used autoprefixed, but the new one did
not. This adds autoprefixer so that `appearance: none` also adds
`-webkit-appearance: none`

This also fixes an issue where there was an `input` inside of a `button`
which made the button extra large. This adds normalizer to ensure CSS
styles are normalized.