Add List-Unsubscribe headers to emails
Adding an unsubscribe link to the emails is great, but what's _really_
great is to add a List-Unsubscribe header. That allows tools that know
to look for this header to do Neat Things like display unsubscribe
buttons in the UI and other cool things. For example, adding this header
makes Constable emails work with https://github.com/pbrisbin/ghu