Fix JS by exporting a function
Closes #229

The new announcement page works if you go directly to it:
https://staging.constable.io/announcements/new, but if you go to it,
visit a different page, and then go back it is not properly set up. This
also happens with the comment form.

The reason is that when you require a module it only requires it once.
So it loads it on the first page load, but then when the page is
re-rendered it doesn't re-require it (which is what we want). So we need
to export a function that we call to set up the event handlers, similar
to what @BlakeWilliams did at first. Anything that should only happen
once should not go in the exported function (stuff like connecting to a
websocket for example), but things like setting up event handlers on the
page should go in the exported function