Display interests for new announcements
Currently, in the daily digest, we're only showing the announcement titles and
users. This is great, but it'd be useful to have more info around _where_ these
new announcements were posted. To help with this, I think it'd be useful to
also display the interests that the new announcement was posted to.