Add image upload for comments and announcements
Uses shubox.io for handling uploads. Each script URL is tied to a domain
so I extracted a config option (:shubox_script_url) that can be changed
with an ENV variable in staging and production.

Because shubox ties each script to a specific host/domain, this feature
won't work for Heroku Review apps since the host is different every
time.