Preview how many users are interested in an announcement
Why?

When writing an announcement, it can be useful to know how many people
it will reach.

How?

This commit adds a preview to the announcement form page that shows how
many users are subscribed to the selected interests. The preview is
updated when interests are added/removed on the announcement form, and
when interests are loaded back into the form from a draft in local
storage.

Closes https://github.com/thoughtbot/constable/issues/135