Preview names of users with selected interests
Why?

When writing an announcement, it can be useful to know specifically who
will receive the announcement because they are subscribed to the
selected interests.

How?

This commit adds a section to the announcement form that displays names
of users who are subscribed to the selected interests. The names are
hidden by default, and can be shown by clicking on the link that shows
how many people are interested. Once the list is expanded, it will
continue to be expanded while adding/removing interests, so you can see
it live-update.

Closes https://github.com/thoughtbot/constable/issues/323