Add service to disable login for inactive users
Currently there is no validation if a user is still an employee or not
which means if they save the session or API token they could potentially
still have access to Constable.

This adds a service to set users as inactive by comparing the user
emails with the emails of active users in hub.

A mix task is also included, `mix constable.disable_inactive_people`