Render previews like the final announcement
Previously we weren't using the same CSS on the markdown preview that we were
using on the final rendered announcement. This led to inconsistency between
the two. Notably, list elements were missing decoration and quote blocks were
rendered differently.

The simplest fix for this is to make sure we apply the same CSS to the
markdown preview that we do to the final announcement. This will help keep
parity between the two and will ensure that as we update the style for one,
the other is automatically updated.