Unify comment/mention emails
Instead of sending a completely different email for when a user is mentioned,
we should instead send a normal comment email but annotate the email to
indicate that the user was mentioned in the body.

I'm not sure if this is the _best_ way to do this, but it felt better to
remove the existing duplication rather than needing to duplicate the entire
template as well. Essentially, `new_comment_mention` becomes an alias for
`new_comment` with the third parameter filled out with the mentioner.