Conditionally show digest sections
Previously, we were always showing each section of the daily digest
(Interests, Announcements, Comments) regardless of if there was content to
show. This was mildly annoying, as it led to empty sections, most commonly
with interests.

To fix this, we can check to see if we have content to show before we actually
show it. This will lead to cleaner emails when there is less content to show.