Add the AGPL license
The intent is to open source this application to serve as a showcase for
our Elixir work and as a general example of a real-world Phoenix
Framework application. We want to avoid a situation where someone else
takes the source code and sells it without making the source available.

The AGPL closes the "service provider" loophole by requiring vendors to
make the source code available if they allow access over a network. This
differs from the GPL, which requires making the source code available
only when the software is distributed.