Use `validate_required` for required inputs
We were previously using a custom presence validator. However, we can
now use `validate_required/2` to do this, which is built in to Ecto.
This also allows us to swith to using `cast/3` rather than `cast/4`,
which means instead of breaking fields into two lists (required and
optional) we present all settable fields for a changeset as one list.

I added `validate_required` after `cast` for any fields which were
user-controllable. I did not add `validate_required` for things like
`account_id` which the user can't act on, should be caught at the
database level, and should frankly not break.

I also removed the calls to `scrub_params` in the controllers, as this
is easy to forget. The combinations of `validate_required` and database
constraints do what we want. We needed to make a single change to
accommodate this, which was to make username generation conditional on a
valid changeset rather than email being truthy.