Add unsubscribe vars to announcement emails
Users shouldn't need to wait for the first comment to unsubscribe from a
thread.

We're actually already adding a link to unsubscribe to new announcements, but
we weren't ever adding the merge vars to fill out the link. This meant that
users would see a template URL instead of the _actual_ URL.

To fix this, we can simply make sure to add these unsubscribe vars to the new
announcements as well as the comments. This means reworking the helper
functions a bit to deal with announcements instead of comments, but those only
_actually_ ever cared about the announcement anyway so the behavior isn't
actually changing.