Update to Elixir 1.3
This is the ground work for updating to Elixir 1.3. We had only a single
warning in our app, but our dependencies had many warnings. I updated as
many of them as I could where the warnings were already addressed and
opened issues on Bamboo, Pact, and Honeybadger to address the rest.

I updated the way the build is run on CI because `--warnings-as-errors`
is not a valid option for `mix test`. We now compile dependencies
(allowing warnings), compile our source (not allowing warnings), and
then test.

Closes #367