Limit width for emails
When a very large image comes in through email, the CSS `max-width` does
not apply in some email clients (Airmail, Gmail, Google Inbox), so the
email appears extremely wide and you must scroll left/right to read each
paragraph.

This is an attempt to limit the email width using a table based layout.
The max width for the New Announcement and New Comment emails should now
be 600px.

See: http://stackoverflow.com/a/24499931