Don't write to localstorage when editing
Currently when editing an announcement the announcement details are
persisted in localStorage. This means when you go back to your draft
announcement it's overwritten with the announcement you last edited.

This changes the JavaScript to check if it's editing an announcement and
to not write to localStorage if you are. It also turns the JavaScript
into a default class export instead of a `setupForm` function to help
keep track of state.