Fix XSS issue with markdown
This fixes an XSS vulnerability anywhere Earmark is used. Earmark
doesn't sanitize user input and lets JavaScript through which we've been
rendering.

To fix this the `HtmlSanitizeEx` package was added and is used to filter
out HTML generated by Earmark.

The JavaScript library `marked` is still vulnerable when writing an
announcement but it's entirely client side.