Adds title to announcement published date
Using the `time_ago_in_words/1` to display the published date is really
nice since it shows something like, `"2 days ago"`. When the date is too
far back, however, it can get confusing knowing when `"45 days ago"` was.

This commit adds the title attribute to the announcement date so that
the announcement published date shows up with the format: `"yyyy-mm-dd"`.