Fix bourbon warnings
Fix bourbon warnings for `display()`, `align-items()`, `animation()`,
and `em()`. All of those functions are now deprecated. Most are
deprecated in favor of using a prefixer library (which constable already
has) and `em()` is no longer suppported. To see a discussion of why it
isn't supported see [this issue](thoughtbot/bourbon#691)