Redesign subscribe and unsubscribe actions
This PR creates a more cohesive and easier to understand subscription component
for users. Previously we had a couple of different styles around this action.
One was under "Manage your interests" and was confusingly similar in visual
style to the new announcement action. The other was on the announcement page and
used hard to understand iconography. It wasn't immediately clear whether or not
someone was subscribed the announcement.

Tangentially this PR unifies the display of # before interests in a couple of
different places. It also cleans up some unnecessary styles by adding a utility
class for a container with padding top.