UI refactor for the announcement page
Goal 1: Create a clear entry point
The previous design lacked a clear entry point into the content. This PR
solves that problem by using a bolder weight and larger size for the
announcement's main heading.

Goal 2: Achieve better typographic hierarchy
The previous type hierarchy was too subtle. This PR solves that by
creating more distinction between the announcement body copy and comment
thread. It also tightens up the line-height and shortens the line-length
on the announcement body. This creates more readable announcements for
Constable users. We've also improved the spacing throughout the app to
feel more solid.

Goal 3: Clarify what subscribing to an announcement means
The old design placed the subscribe button adjacent to the
announcement's tags. This creates confusion around what the reader is
subscribing to. Is is the message or the tags? Is it all of the tags or
just one? To fix this we've moved the subscribe button below the title
and adjacent to the author's name. We've also clarified the button's
label to be more specific. It now says "Subscribe to thread."

Goal 4: Increase Constable's friendliness
The old design had avatars that were very small. Because of that each
one looked more-or-less the same. This PR increases both an announcement
author's avatar and a commenter's. This helps Constable readers
recognize faces better.