Use CircleCi 2
What changed?
=============

We've been having issues running our test suite in CircleCi. The version
of asdf we were using was pretty old. Instead of investing time fixing
that, it seemed like a good idea to just go ahead and switch to
CircleCi 2.

This removes the old circle config file and adds a new one for circle ci
2.