MVP UI/UX refactor for subscribed interests bar
This PR displays the edit your interests feature conditionally. We have
a flag @show_all that allows us to display interests in
announcements/index.html.eex page. This UI refactor makes the experience
of viewing `All announcements` cleaner and more focused. It does the
same for `Your interests` because the action of editing is now tied only
to that tab.

Co-authored-by: German Velasco <germsvel@gmail.com>