Header and announcement list UI refactors
This PR is concerned mostly with Constable's header. It refactors the UI
for more order. It gives the search feature more prominence to help
readers find announcements easier. It also tightens up the vertical
space previously used by the widget to access one's subscribed
interests. Now that anchors the right side of the navigation container.
Lastly, it reworks the small screen header design so everything has
ample living space. Previously, the elements felt squished together on
small screens.

In addition this PR makes a few changes to the announcement list. First,
it scales back the use of red for the announcement list headlines. Red
is a powerful color and we want to reserve that power for things like
buttons, inline links or other strong actions. Lastly, it makes the
avatars a little smaller and increases the size of the headlines for
improved hierarchy.