Compile assets directly on mix test
When we moved assets from Phoenix 1.2 to 1.3 locations under `assets/`,
we updated our CI config and README to run `bin/test_suite` when running
the entire test suite because we had to ensure we were compiling assets
before acceptance specs.

I found a better way to do that without having to resort to
`bin/test_suite`. We can do it from within the mix file, and this just
updates that to do it within `mix test`.