Use Phoenix.Param protocol for path_helpers
The Phoenix.Param protocol takes care of retrieving the id from an
announcement when passed into a path helper.

So instead of doing,

```elixir
announcement_path(@announcement.id)
```

We can simply do,

```elixir
announcement_path(@announcement)
```

This is beneficial for us because we're hoping to modify the protocol
for announcements so that it uses slugs in the future.