Add slugs to announcements
What?
=====

We add slugs to announcements for better urls. We do so in a way that
maintains backwards compatibility and allows for urls with ids and urls
with the new slugs.

A note on terminology
=====================

We technically add two things that could be called "slugs". One is a
column in the announcements table. The other is what is now at the end
of an announcement path (e.g. "announcements/:slug").

The slugs in the database are a slugified version of the announcement's
title. The slugs in the path are really a composite slug of the
announcement id and the slug in the database field. So the slugs in the
url have the form of `id-slugified-title'.

Details
========

We add slugs to announcements paths that have an `id-sluggified-title`
format. We do this by implementing the Phoenix.Param protocol for
announcement so that all path helpers automatically generate a slug that
is a combination of the id and the slug.

We also add a columng "slug" to announcements table. We are not
including an index because we do not intend to retrieve records by the
slug. We'll retrieve them by id.

In order to persistent and associations easily with the new slugs, we
use a custom ecto type that knows how to cast `id-sluggified-title`
formats as well as simple integers. We set that type as the primary key
in announcement so that ecto knows how to handle it.

Why this approach?
==================

The benefit of this approach is that we don't have to worry about what
form the "id" will come in. It could be an integer or it could be the
new slugs. It'll cast it when saving the id to the database and
essentially throw away the slug (for the purposes of the primary id).
This means that all old urls are automatically okay, while new urls will
be supported. It also means that changing the title of the announcement
will break any old links someone saved outside of constable.