Fix intermittent failure due to sort order
The api searches controller test was failing intermittently because the
test compares a list of announcements with those returns from the
search. Since there was no order specified in the query, the comparison
would fail if the order was different even if all the correct elements
were returned.

In other words,

```
assert [a, b] == [b, a] # => false
```

We add a `order_by` to the `Announcement.search` so that we guarantee an
order.