Remove CLIENT_ variables from .sample.env
What?
=====
We remove `CLIENT_ID` and `CLIENT_SECRET` from `.sample.env`.

Why?
====

When someone clones the repo for the first time and runs `bin/setup`
they cannot log in into the application because they do not have
`CLIENT_ID` and `CLIENT_SECRET` set.

Those variables should be automatically set in `bin/setup`, but the
problem is that we copy `.sample.env` with default values for those two
variables, and at the same time we check the `.env` file for `CLIENT_ID`
and `CLIENT_SECRET` to make sure we're not overriding keys.

The result is that we copy `.sample.env` with the default keys and then
at the end of `bin/setup` we don't actually copy the staging keys
because we already have some defined (the default ones).