Return last discussed announcements first
A previous commit c7cec8962f6b9c7f08b8b54f26991eed26f79fcb introduced an
ordering of announcements for `Announcement.search/3` in order to fix an
intermittently failing test. But that also introduced a bug. The order
of announcements returned from the search were no longer ordered with the
`Announcement.last_discussed_first/1`. We fix that here by removing the
ordering on the `search/3` function and relying on
`last_discussed_first/1` for the search controllers, both for the web
and the api.