Do not show deactivated users in recipients preview
When creating an announcement, deactivated users still show up in the
list of interested users for the `interest`s associated to the
announcement. This commit scopes the query so that only active users are
shown there.

Other changes:

- We add a test for the AnnouncementCreator module to ensure (and
document) that deactivated users subscribed to an `interest` do not
receive an email when the announcement is published.