Update Wallaby to 0.20.x
Our version of Wallaby is fairly outdated. We update to the newest
version `0.20.0`.

The biggest changes that we made include the `Wallaby.Query` module.
This module exposes functions like `link/1`, `button/1`, etc. that are
meant to play nicely with `Wallaby.Browser` functions, such as
`click/2`.

So we can do things like,

```elixir
session
|> fill_in(text_field("Name"), with: "thoughtbot")
|> click(button("Submit"))
```

where `fill_in/3` and `click/2` come from the browser module and
`text_field/1` and `button/1` come form the query module.

Other changes:

The newest version of Wallaby also re-raises any javascript errors. For
the longest time, `Shubox` has been undefined in testing, but now that
Wallaby re-raises that error we had to handle it. I simply check to see
if it's defined since the constant `Shubox` comes in through a script
included in a header.