Update phoenix_html to version 2.12.0
The library changed how the `link` method works. Previously there was a
surrounding "parent form" generated which we hooked into with our custom
turbolinks/html JS. The phoenix JS has changed to be more sophisticated and now
generated a hidden form which is used to submit links for non-GET actions.

As a result of this, our turbolinks custom stuff stopped working, and I decided
to remove it instead of change it to attempt to keep pace with the framework
changed JS.