Use UTC DateTime in schema (instead of NaiveDateTime)
Ecto changed how it handles timezones from 1.x to 2.x, with 2.x no longer
assuming there was a timezone for a timestamp, and returning a NaiveDateTime
instead of an elixir DateTime or an Ecto.DateTime.

A side effect of this is that API responses from Constable suddenly stopped
including the timezone portion of the ISO8601 format timestamp response that
most clients assume would be available.

This change tells Ecto models to use a utc datetime type and restores the
inclusion of timezone in the ISO8601 values from the API (and elsewhere).