Allow HTML5 in announcements
What?
======

We change `HtmlSanitizeEx` from `basic_html` to `html5`. That function
is still a work-in-progress according to some comments in the source
code, but I think it is safe enough for us to use since this is an
internal tool. See source code at:
https://github.com/rrrene/html_sanitize_ex/blob/master/lib/html_sanitize_ex/scrubber/html5.ex

Why?
=====

The main driver behind this is the ability to add <iframe>s so we can
add video in an announcement. The `html_sanitize_ex` library allows us
to create a custom sanitizer where we could do basic html + iframes, but
I think it makes sense to go with the html5 sanitizer even if it's a
work in progress.

At some point, it might be nice to add oembed so that someone can just
put a link and we can retrieve it for them, but allowing the rendering
of <iframe>s is a pre-requisite for that anyway, so I think this is a
good first step for those that want to add video to their announcements.