Update incorrect shubox script URL in dev environment
The URL value we were using was outdated (and just erroring in local
development), and needed to be replaced with a new value.

Separately, since this project is open source, it's sort of awkward to expose
the actual value for the development URL of an account TB owns but does not
necessarily want to share with other contributors.

I've added the thougthbot constable shubox account credentials to the team
dropbox so that anyone setting up an environment can add that to their local
.env and have uploads working locally, if needed.