Support emoji in announcement titles
We were seeing app errors occur when people tried to create announcements where
the title contained an Emoji.

Example: https://app.honeybadger.io/projects/48229/faults/44121036#notice-summary

We tracked this back to the Announcement changeset massaging a `slug` value out
of the announcement title while it saves. We use the Slugger library to do this,
which has a bug in version 0.3.0 around removing chars outside the expected
range: https://github.com/h4cc/slugger/issues/32

A side effect of this was that our title values wound up with non-UTF8 chars in
them and postgres threw an error.

This change locks the version of Slugger to one which was working, and adds test
coverage for including Emoji in announcement titles.