Set json serializer for OAuth2 client
When we upgraded the oauth2 library to v1.0.x, there was an incompatible
changes that we did not fulfill. The change required us to specify the
Json parser we want to use. We do that in this commit.

Looking at our `mix.exs` file, it seems we have both Poison and Jason,
libraries that serve the same purpose, so it seems like we should choose
one in the future. For now, I selected Jason since that is the faster of
the two and likely the one we will migrate everything else to.