Make user names clickable to see their announcements
This leverages the `/announcements#index` route with the query params
`user_id` that we introduced in order to display a user's own
announcements. We use it to display the list of announcements for a
any user when we click on their name.

The only strange thing about this is that we leave the nav bar in a
completely unselected state, since seeing another user's announcements
does not have an equivalent "tab" as we might have for "All
announcements" or "Your announcements".