Adds "Your Announcements" to nav bar (#613)
What changed?
------------

This commit adds a new tab into the announcements-interests nav bar. The
tab, titled "Your announcements", allows users to see announcements they
have made.

Why?
----

When users make announcements, they have the ability to edit them in the
future. Users may also want to see comments people have left in one of
their previous announcements. But I have always found it difficult to
find my previous announcements.

Comments on some of the code
----------------------------

Most of the implementation was included alongside the implementation of
the other two `announcement#index` actions, leveraging pattern matching
to filter things correctly. It seems like there's a lot of logic in the
controller, and it would be good to refactor it.
I have decided not to do more of that in this PR to avoid introducing
more changes.

There are also some unrelated changes in files I have touched due to the
elixir formatter. I have decided to include those here rather than
turning the formatter off, but it may be good to run the formatter
across all files in a single PR so that we do not have constant PRs that
have a lot of unrelated changes due to auto-formatting.