Render new article markdown preview via LiveView
We get the preview using live view (currently under
announcements/new?live=true) to render correctly. The title and body
render fine.

In order to handle selectize (on interests), we use live view hooks to
reinitialize selectize every time live view mounts or updates.

We keep the organization of /hook files, but we extract the live socket
setup to app.js because the previous organization had broken the live
socket connection. We're also relying on `import` from within the js
files instead of `global.constable`.

Note on CSRF token:
-------------------

We pass the csrf token from the controller to the template so that
submitting the live view form to the regular controller works.

Note: it seems that Plug ~> 1.8 will fix this, since it'll allow for
dumping the csrf token from one process into another one. That'll
probably come in Phoenix 1.5 since it'll require a bump in Plug. In the
meantime, this workaround works fine OR we could just do a phx-submit
event on the form and handle that event in AnnouncementLive.

Co-authored-by: Wil Hall <wil@wilhall.com>
Co-authored-by: Alejandro Dustet <aledustet@gmail.com>