Fix Settings page, remove Settings modal
This fixes the rendering of the Settings page, which was not rendered
using a layout and therefore lacked CSS styling.

In conjunction, we remove the modal UI from the viewing of Settings and
instead leverage a standard page view. Modals often pose accessibility
issues (ours did) and it meant that did not have a URL to point at for
settings (our emails point to a `/settings` URL, though).

Fixes: https://github.com/thoughtbot/constable/issues/700