Consistently render Interests as badges
We've long rendered Interest as a "badge" on the Announcements index
and show pages. In [281f3f7] we tweaked badges and started using the
Badge component from tbds.

However, on Interest index and show pages, we were not rendering them
visually as a "badge," rather they just rendered as plain text.

This changes that to carry the visual representation of an Interest to
be a "badge" throughout Constable, creating a consistent visual system.

[281f3f7]: https://github.com/thoughtbot/constable/commit/281f3f7429c18106ba96c3bf92dc558e26d19301