Refactor media queries and breakpoints
In [438cfbb] we removed Neat (v1), but copied over its `media` mixin
because, at the time, it was used quite a bit and would've been more
complicated to pull out. That situation has changed and we're now in a
better position to remove it.

The `media` mixin we copied over is not a great mixin because it uses
string parsing instead of arguments. This has perhaps led us to use a
mix of `min-width` and `max-width` queries. It made our Sass overly
complex and hard to follow.

This commit:

- Removes the `media` mixin in favor of using vanilla CSS
  `@media` queries
- Moves everything over to `min-width` queries (commonly referred to as
  "mobile-first")

[438cfbb]: https://github.com/thoughtbot/constable/commit/438cfbbf13f8f147649676037244b52d0d351be9