Show one announcement per comment in "Your Comments"
What changed?
=============

Prior to this change, if someone commented multiple times on the same
announcement, clicking on "Your Comments" would show them multiple
entries for the same announcement (one entry per comment).

We now include a DISTINCT on the query to only show one announcement in
"Your Comments" tab regardless of how many times the user has commented
on it.