Add "View on Constable" link to top of emails (#826)
We add a "View on Constable" link to the top of the new announcement and
new comment emails. Because the announcements and comments can vary in
length, it is sometimes difficult to find the link if the reader prefers
to see the announcement in the browser.

This brings a gray link to the top that is same color and size as the
interests displayed on the top left.