Fetch Profiles info rather than memoize calls (#891)
What?
=====

We make a change in how and when we fetch profile information for users
(profile_url and profile_image_url). Currently in production we memoize
the calls to hub when we're trying to load a user's profile image url or
profile image. The problem with that is that if the cache isn't warm, we
have to make N queries to hub trying to fetch every user's image
url when loading a list of announcements, where N is the number of users
that have made a comment, made an announcement, plug the current user.
That makes for a very slow page load, which sometimes times out.

In order to get around that, we now store the profile_url and
profile_image_url as part of the users table. This gets rid of the N
queries at load time, and we perform those actions when (a) a user is
created, (b) a user is reactivated, or (c) daily via a mix task (the mix
task only updates users who do not have a url set).

When a user is created or reactivated, we fetch the profile information
asynchronously. To achieve this, we add a `Constable.TaskSupervisor`
which can handle "one off" tasks.

The mix task, on the other hand, performs a "brute force" approach to
fetching the image urls, making N requests for N users. But since we're
doing it on a background job, this seems to be fine for now.

Finally, an even better approach that can be taken in the future (though
it requires changes in hub) would be for hub to provide the information
to us on user creation. I believe hub is already the source of truth for
the email and name, and it provides those to constable on user creation
(see api/user_controller#create). If it could also provide the
profile_url and profile_image_url at that point, we wouldn't have to
fetch them afterwards.

Currently, we roughly have this workflow on user creation:

```
hub | -- create a user (email, name)       --> | constable
hub | <-- what is the user's profile info? --  | constable
hub | -- profile_url and profile_image_url --> | constable
```

If hub could provide the information needed upon creation, we would
simplify the workflow to:

```
hub | -- create a user (email, name, profile fields) --> | constable
```