Only render link to profile url if available
In the previous commit we changed how we fetch profile urls from hub.
Instead of doing it on demand when a page is loading, we now do it
asynchronously. One of the problems with that is that when the
`user.profile_url` is `nil` (because we haven't yet fetched it), the
user's settings page breaks because we're passing `nil` to Phoenix's
`link` function.