Load shubox script asynchronously (#912)
From time to time, Constable seems completely unresponsive. When looking
into it, George and Adam, discovered that the shubox script is not
loading. Looking at shubox's docs, it seems like we might have an old
way of dealing with this. We'll update that in a future commit. For now,
we just load it asynchronously so as to not block the loading of the
page.

Shubox docs: https://github.com/shuboxio/shubox.js/blob/master/README.md